requirements.txt
trio==0.22.0
grpcio==1.51.1
grpcio-tools==1.51.1
requests==2.28.1
psycopg2-binary==2.9.6
